+++ adminer/script.inc.php
@@ -6,7 +6,7 @@ header("Content-Type: application/json; charset=utf-8");
 if ($_GET["script"] == "db") {
 	$sums = array("Data_length" => 0, "Index_length" => 0, "Data_free" => 0);
 	foreach (table_status() as $name => $table_status) {
-		json_row("Comment-$name", h($table_status["Comment"]));
+		json_row("Comment-$name", h($table_status["Comment"]) . ($table_status["Error"] ? " <span class='error'>" . h($table_status["Error"]) . "</span>" : ""));
 		if (!is_view($table_status) || preg_match('~materialized~i', $table_status["Engine"])) {
 			foreach (array("Engine", "Collation") as $key) {
 				json_row("$key-$name", h($table_status[$key]));